Echo status in worker nodes/Case when master fails then back up and active_workers exist

+++ files/root/exareme/bootstrap.sh
@@ -43,19 +43,30 @@ if [ -z ${CONSULURL} ]; then echo "CONSULURL is unset"; exit; fi
            SPACE=' '
            . /root/exareme/start_worker.sh
            if [ "$(curl -o -i -s -w "%{http_code}\n" ${CONSULURL}/v1/kv/${EXAREME_ACTIVE_WORKERS_PATH}/{$NODE_NAME}?keys)" = "200" ]; then
-                ssh -oStrictHostKeyChecking=no $SH """sed -i  "/`echo $NODE_NAME`/d" /root/exareme/etc/exareme/workers"""
+                ssh -oStrictHostKeyChecking=no $SH """sed -i  "/`echo $NODE_NAME`/d" /root/exareme/etc/exareme/workers"""       #sed -i == delete line from etc/exareme/workers
            fi
            curl -X PUT -d @- $CONSULURL/v1/kv/$EXAREME_ACTIVE_WORKERS_PATH/$NODE_NAME <<< $IP
            echo $IP$SPACE$NODE_NAME | ssh -oStrictHostKeyChecking=no $SH "cat >> /root/exareme/etc/exareme/workers"     #write workers's IP into master's worker file
-           while [ ! -f "/tmp/exareme/var/log/exareme-*.log" ]; do
+           NAME=$(curl -s $CONSULURL/v1/kv/$EXAREME_MASTER_PATH/?keys | jq -r '.[]' | sed "s/$EXAREME_MASTER_PATH\///g")
+           while [ ! -f "/tmp/exareme/var/log/exareme-worker.log" ]; do
+                echo "Trying to connect to master with IP "$SH" and NAME "$NAME"."
                 sleep 2
            done
+           echo -e "\nConnected to master with IP "$SH" and NAME "$NAME"."
         else    #the system just created
            MY_OLIP=$(/sbin/ifconfig $1 | grep "inet " | awk -F: '{print $2}' | grep '10.20' | awk '{print $1;}' | head -n 1)
            curl -X PUT -d @- $CONSULURL/v1/kv/$EXAREME_WORKERS_PATH/$MY_OLIP <<< $NODE_NAME
-           while [ ! -f "/tmp/exareme/var/log/exareme-*.log" ]; do
+           while [ ! -f "/tmp/exareme/var/log/exareme-worker.log" ] ; do
+                while [ "$(curl -o -i -s -w "%{http_code}\n" ${CONSULURL}/v1/kv/${EXAREME_ACTIVE_WORKERS_PATH}/?keys)" != "200" ] ; do
+                    sleep 1
+                done
+                SH=$(curl -s $CONSULURL/v1/kv/$EXAREME_MASTER_PATH/$(curl -s $CONSULURL/v1/kv/$EXAREME_MASTER_PATH/?keys | jq -r '.[]' | sed "s/$EXAREME_MASTER_PATH\///g")?raw)
+                NAME=$(curl -s $CONSULURL/v1/kv/$EXAREME_MASTER_PATH/?keys | jq -r '.[]' | sed "s/$EXAREME_MASTER_PATH\///g")
+                echo "Trying to connect to master with IP "$SH" and NAME "$NAME"..."
                 sleep 2
            done
+           NAME=$(curl -s $CONSULURL/v1/kv/$EXAREME_MASTER_PATH/?keys | jq -r '.[]' | sed "s/$EXAREME_MASTER_PATH\///g")
+           echo -e "\nConnected to master with IP "$SH" and NAME "$NAME"."
 	    fi
     else #this is the master
         while [ "$(curl -s ${CONSULURL}/v1/health/state/passing | jq -r '.[].Status')" != "passing" ]; do			#sleep 2
@@ -64,26 +75,41 @@ if [ -z ${CONSULURL} ]; then echo "CONSULURL is unset"; exit; fi
         echo -n $NODE_NAME > /root/exareme/etc/exareme/name
         /sbin/ifconfig $1 | grep "inet " | awk -F: '{print $2}' | grep '10.20' | awk '{print $1;}' | head -n 1 > /root/exareme/etc/exareme/master
         MY_OLIP=$(/sbin/ifconfig $1 | grep "inet " | awk -F: '{print $2}' | grep '10.20' | awk '{print $1;}' | head -n 1)
-        curl -X PUT -d @- $CONSULURL/v1/kv/$EXAREME_MASTER_PATH/$NODE_NAME <<< $MY_OLIP
-        WORKERS_UP=0
-        while [ $WORKERS_UP != $EXA_WORKERS_WAIT ]; do		#for test $EXA_WORKERS_WAIT ==1
-            sleep 2
-            curl -s $CONSULURL/v1/kv/$EXAREME_WORKERS_PATH/?keys | jq -r '.[]' | sed "s/$EXAREME_WORKERS_PATH\///g"  \
-            | head -n $EXA_WORKERS_WAIT > /root/exareme/etc/exareme/workers
-                WORKERS_UP=`cat etc/exareme/workers | wc -l`
-                echo "Waiting for " $((EXA_WORKERS_WAIT-WORKERS_UP)) " more exareme workers..."
-        done
-        SPACE=' '
-        for i in `cat /root/exareme/etc/exareme/workers` ; do
-            ssh -oStrictHostKeyChecking=no $i date
-            curl -X PUT -d @- $CONSULURL/v1/kv/$EXAREME_ACTIVE_WORKERS_PATH/$(curl -s $CONSULURL/v1/kv/$EXAREME_WORKERS_PATH/$i?raw) <<< $i
-            WORKER_NAME=$(curl -s $CONSULURL/v1/kv/$EXAREME_WORKERS_PATH/$i?raw)
-            sed -i "/$i/c`echo $i$SPACE$WORKER_NAME`" /root/exareme/etc/exareme/workers
-            curl -X DELETE $CONSULURL/v1/kv/$EXAREME_WORKERS_PATH/$i
-        done
-        ./bin/exareme-admin.sh --update
-        sleep 3
-        ./bin/exareme-admin.sh --start
+        if [ "$(curl -o -i -s -w "%{http_code}\n" ${CONSULURL}/v1/kv/${EXAREME_MASTER_PATH}/?keys)" = "200" ]; then
+            if [ "$(curl -o -i -s -w "%{http_code}\n" ${CONSULURL}/v1/kv/${EXAREME_ACTIVE_WORKERS_PATH}/?keys)" = "200" ]; then
+                for i in `curl -s $CONSULURL/v1/kv/${EXAREME_ACTIVE_WORKERS_PATH}/?keys | jq -r '.[]' | sed "s/${EXAREME_ACTIVE_WORKERS_PATH}\///g"` ; do
+                    IP=$(curl -s $CONSULURL/v1/kv/${EXAREME_ACTIVE_WORKERS_PATH}/$i?raw)
+                    SPACE=' '
+                    echo $IP$SPACE$i >> /root/exareme/etc/exareme/workers
+                    ssh -oStrictHostKeyChecking=no $IP date
+                done
+                ./bin/exareme-admin.sh --stop
+                sleep 3
+                ./bin/exareme-admin.sh --start
+            #else mipws oi workers den exoun prolavei na einai active kai einai available?
+            fi
+        else
+            curl -X PUT -d @- $CONSULURL/v1/kv/$EXAREME_MASTER_PATH/$NODE_NAME <<< $MY_OLIP
+            WORKERS_UP=0
+            while [ $WORKERS_UP != $EXA_WORKERS_WAIT ]; do		#for test $EXA_WORKERS_WAIT ==1
+                sleep 2
+                curl -s $CONSULURL/v1/kv/$EXAREME_WORKERS_PATH/?keys | jq -r '.[]' | sed "s/$EXAREME_WORKERS_PATH\///g"  \
+                | head -n $EXA_WORKERS_WAIT > /root/exareme/etc/exareme/workers
+                    WORKERS_UP=`cat etc/exareme/workers | wc -l`
+                    echo "Waiting for " $((EXA_WORKERS_WAIT-WORKERS_UP)) " more exareme workers..."
+            done
+            SPACE=' '
+            for i in `cat /root/exareme/etc/exareme/workers` ; do
+                ssh -oStrictHostKeyChecking=no $i date
+                curl -X PUT -d @- $CONSULURL/v1/kv/$EXAREME_ACTIVE_WORKERS_PATH/$(curl -s $CONSULURL/v1/kv/$EXAREME_WORKERS_PATH/$i?raw) <<< $i
+                WORKER_NAME=$(curl -s $CONSULURL/v1/kv/$EXAREME_WORKERS_PATH/$i?raw)
+                sed -i "/$i/c`echo $i$SPACE$WORKER_NAME`" /root/exareme/etc/exareme/workers
+                curl -X DELETE $CONSULURL/v1/kv/$EXAREME_WORKERS_PATH/$i
+            done
+            ./bin/exareme-admin.sh --update
+            sleep 3
+            ./bin/exareme-admin.sh --start
+        fi
 fi
 
 if [ -e "/tmp/exareme/var/log/exareme-*.log" ]
